create TestHelper.ts to isolate variables update imports
